COMMERCE-6525 Refactor move sitemap providers to util directory
