+++ modules/apps/commerce/commerce-product-content-web/src/main/java/com/liferay/commerce/product/content/web/internal/util/CPDefinitionSitemapURLProvider.java
@@ -12,34 +12,40 @@
  * details.
  */
 
-package com.liferay.commerce.product.content.web.internal.sitemap;
+package com.liferay.commerce.product.content.web.internal.util;
 
-import com.liferay.asset.kernel.model.AssetCategory;
-import com.liferay.asset.kernel.model.AssetVocabulary;
-import com.liferay.asset.kernel.service.AssetCategoryService;
-import com.liferay.asset.kernel.service.AssetVocabularyService;
+import com.liferay.commerce.account.model.CommerceAccount;
+import com.liferay.commerce.account.util.CommerceAccountHelper;
+import com.liferay.commerce.product.catalog.CPCatalogEntry;
+import com.liferay.commerce.product.catalog.CPQuery;
 import com.liferay.commerce.product.constants.CPPortletKeys;
+import com.liferay.commerce.product.data.source.CPDataSourceResult;
+import com.liferay.commerce.product.model.CPDefinition;
+import com.liferay.commerce.product.model.CProduct;
+import com.liferay.commerce.product.service.CPDefinitionLocalService;
+import com.liferay.commerce.product.service.CommerceChannelLocalService;
 import com.liferay.commerce.product.url.CPFriendlyURL;
+import com.liferay.commerce.product.util.CPDefinitionHelper;
 import com.liferay.friendly.url.model.FriendlyURLEntry;
 import com.liferay.friendly.url.service.FriendlyURLEntryLocalService;
 import com.liferay.layout.admin.kernel.model.LayoutTypePortletConstants;
 import com.liferay.layout.admin.kernel.util.Sitemap;
 import com.liferay.layout.admin.kernel.util.SitemapURLProvider;
-import com.liferay.portal.kernel.dao.orm.QueryUtil;
 import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.model.Company;
-import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.model.Layout;
 import com.liferay.portal.kernel.model.LayoutSet;
-import com.liferay.portal.kernel.service.CompanyLocalService;
+import com.liferay.portal.kernel.search.Field;
+import com.liferay.portal.kernel.search.SearchContext;
 import com.liferay.portal.kernel.service.LayoutLocalService;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.GetterUtil;
+import com.liferay.portal.kernel.util.HashMapBuilder;
 import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.UnicodeProperties;
+import com.liferay.portal.kernel.workflow.WorkflowConstants;
 import com.liferay.portal.kernel.xml.Element;
 
-import java.util.List;
+import java.io.Serializable;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -50,11 +56,11 @@ import org.osgi.service.component.annotations.Reference;
 @Component(
 	enabled = false, immediate = true, service = SitemapURLProvider.class
 )
-public class AssetCategorySitemapURLProvider implements SitemapURLProvider {
+public class CPDefinitionSitemapURLProvider implements SitemapURLProvider {
 
 	@Override
 	public String getClassName() {
-		return AssetCategory.class.getName();
+		return CPDefinition.class.getName();
 	}
 
 	@Override
@@ -65,34 +71,51 @@ public class AssetCategorySitemapURLProvider implements SitemapURLProvider {
 
 		long plid = _portal.getPlidFromPortletId(
 			layoutSet.getGroupId(), layoutSet.isPrivateLayout(),
-			CPPortletKeys.CP_CATEGORY_CONTENT_WEB);
+			CPPortletKeys.CP_CONTENT_WEB);
 
 		Layout layout = _layoutLocalService.getLayout(plid);
 
 		if (layoutUuid.equals(layout.getUuid())) {
-			Group group = layoutSet.getGroup();
+			long groupId =
+				_commerceChannelLocalService.
+					getCommerceChannelGroupIdBySiteGroupId(
+						layoutSet.getGroupId());
 
-			Company company = _companyLocalService.getCompany(
-				group.getCompanyId());
+			CommerceAccount commerceAccount =
+				_commerceAccountHelper.getCurrentCommerceAccount(
+					themeDisplay.getRequest());
 
-			List<AssetVocabulary> assetVocabularies =
-				_assetVocabularyService.getGroupVocabularies(
-					company.getGroupId(),
-					group.getName(themeDisplay.getLocale()), QueryUtil.ALL_POS,
-					QueryUtil.ALL_POS, null);
+			SearchContext searchContext = new SearchContext();
 
-			if (assetVocabularies.size() == 1) {
-				AssetVocabulary assetVocabulary = assetVocabularies.get(0);
+			searchContext.setAttributes(
+				HashMapBuilder.<String, Serializable>put(
+					Field.STATUS, WorkflowConstants.STATUS_APPROVED
+				).put(
+					"commerceAccountGroupIds",
+					_commerceAccountHelper.getCommerceAccountGroupIds(
+						commerceAccount.getCommerceAccountId())
+				).put(
+					"commerceChannelGroupId", groupId
+				).build());
 
-				List<AssetCategory> assetCategories =
-					_assetCategoryService.getVocabularyRootCategories(
-						assetVocabulary.getGroupId(),
-						assetVocabulary.getVocabularyId(), QueryUtil.ALL_POS,
-						QueryUtil.ALL_POS, null);
+			searchContext.setCompanyId(themeDisplay.getCompanyId());
 
-				for (AssetCategory assetCategory : assetCategories) {
+			CPQuery cpQuery = new CPQuery();
+
+			cpQuery.setOrderByCol1("title");
+			cpQuery.setOrderByCol2("modifiedDate");
+			cpQuery.setOrderByType1("ASC");
+			cpQuery.setOrderByType2("DESC");
+
+			CPDataSourceResult cpDataSourceResult = _cpDefinitionHelper.search(
+				groupId, searchContext, cpQuery, -1, -1);
+
+			for (CPCatalogEntry cpCatalogEntry :
+					cpDataSourceResult.getCPCatalogEntries()) {
+
+				if (layoutUuid.equals(layout.getUuid())) {
 					visitLayout(
-						element, layout, assetCategory.getCategoryId(),
+						element, layout, cpCatalogEntry.getCPDefinitionId(),
 						themeDisplay);
 				}
 			}
@@ -106,7 +129,7 @@ public class AssetCategorySitemapURLProvider implements SitemapURLProvider {
 	}
 
 	protected void visitLayout(
-			Element element, Layout layout, long assetCategoryId,
+			Element element, Layout layout, long cpDefinitionId,
 			ThemeDisplay themeDisplay)
 		throws PortalException {
 
@@ -128,12 +151,16 @@ public class AssetCategorySitemapURLProvider implements SitemapURLProvider {
 		String currentSiteURL = _portal.getGroupFriendlyURL(
 			layout.getLayoutSet(), themeDisplay);
 
-		String urlSeparator = _cpFriendlyURL.getAssetCategoryURLSeparator(
+		String urlSeparator = _cpFriendlyURL.getProductURLSeparator(
 			themeDisplay.getCompanyId());
+
+		CPDefinition cpDefinition = _cpDefinitionLocalService.getCPDefinition(
+			cpDefinitionId);
 
 		FriendlyURLEntry friendlyURLEntry =
 			_friendlyURLEntryLocalService.getMainFriendlyURLEntry(
-				_portal.getClassNameId(AssetCategory.class), assetCategoryId);
+				_portal.getClassNameId(CProduct.class),
+				cpDefinition.getCProductId());
 
 		String productFriendlyURL =
 			currentSiteURL + urlSeparator +
@@ -145,13 +172,16 @@ public class AssetCategorySitemapURLProvider implements SitemapURLProvider {
 	}
 
 	@Reference
-	private AssetCategoryService _assetCategoryService;
+	private CommerceAccountHelper _commerceAccountHelper;
 
 	@Reference
-	private AssetVocabularyService _assetVocabularyService;
+	private CommerceChannelLocalService _commerceChannelLocalService;
 
 	@Reference
-	private CompanyLocalService _companyLocalService;
+	private CPDefinitionHelper _cpDefinitionHelper;
+
+	@Reference
+	private CPDefinitionLocalService _cpDefinitionLocalService;
 
 	@Reference
 	private CPFriendlyURL _cpFriendlyURL;
